requirements.txt
colorama==0.4.6
dnspython==2.7.0
python-dateutil==2.9.0.post0
python-nmap==0.7.1
python-whois==0.9.5
six==1.17.0
